Increasing cache timeouts and change RSS guid
GUID is a unique identifier in an RSS feed. For a given rental item there
are three ids: ID, CID, and inv. ID changes each time a movie is rented AND
returned. CID is unique to each registered credit card. inv only changes
when a new rental happens. This makes it the most ideal for an RSS feed.